Require Ruby >= 1.9.3
The gemspec requires activemodel, but does not specify a version
dependency. However, Rails 4 now requires >= 1.9.3. This means that it
is impossible to run `bundle` on 1.9.2 and REE, which means it's
impossible to develop on those versions. This is also causing tests to
not run on Travis.

Unfortunately, there isn't a good way to fix this in the gemspec, since
it's a bad idea to have conditional dependencies. While I'm sure some
people are still on 1.9.2, it is obsolete and unsupported. So the best
way to solve this is just to drop support.

This means we also no longer need to use custom Gemfiles on Travis.
These Gemfiles ensured that anything before 1.9.3 set the ActiveModel
dependency to 3.x and anything after was using 4.x. This is false
because this is inconsistent with what the gemspec is saying, which is
that quickeebooks depends on *any* version of ActiveModel (meaning that
people who install this gem on <= 1.9.3 will actually always get
ActiveModel 4, not 3).